fix(lint): resolve eslint errors in peer-note fetch UI
- Move isPeerContentFailure/isPeerContentReady out of PeerNoteFallback into peerNoteStatus.ts (react-refresh/only-export-components)
- usePeerNoteContent: store only the resolved result (keyed by request) and derive 'idle'/'loading', so the effect never calls setState synchronously (react-hooks/set-state-in-effect)
