+++ src/components/editor/PeerNoteFallback.tsx
@@ -1,23 +1,7 @@
 import { Loader2 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import type { PeerNoteContentStatus } from '@/hooks/usePeerNoteContent';
-
-const MESSAGES: Partial<Record<PeerNoteContentStatus, string>> = {
-    offline: "Can't reach the author to load this shared note. They may be offline.",
-    forbidden: 'You no longer have access to this note.',
-    notfound: 'This note no longer exists.',
-    error: 'Something went wrong loading this shared note.',
-};
-
-export function isPeerContentFailure(status: PeerNoteContentStatus): boolean {
-    return status in MESSAGES;
-}
-
-const READY_STATUSES: PeerNoteContentStatus[] = ['local', 'fetched', 'empty'];
-
-export function isPeerContentReady(status: PeerNoteContentStatus): boolean {
-    return READY_STATUSES.includes(status);
-}
+import { PEER_NOTE_MESSAGES } from './peerNoteStatus';
 
 export function PeerNoteLoading() {
     return (
@@ -39,7 +23,7 @@ export function PeerNoteError({
 }) {
     return (
         <div className="flex flex-col items-center justify-center h-full text-muted-foreground gap-3 px-6 text-center">
-            <p className="text-sm font-medium">{MESSAGES[status] ?? MESSAGES.error}</p>
+            <p className="text-sm font-medium">{PEER_NOTE_MESSAGES[status] ?? PEER_NOTE_MESSAGES.error}</p>
             <div className="flex gap-2">
                 <Button variant="outline" size="sm" onClick={onRetry}>Retry</Button>
                 <Button variant="ghost" size="sm" onClick={onBack}>Back to list</Button>
